Hotfix/0.2.0

+++ models/document/types.go
@@ -4,7 +4,7 @@ type Coins []Coin
 
 type Coin struct {
 	Denom  string `bson:"denom"`
-	Amount int64  `bson:"amount"`
+	Amount float64  `bson:"amount"`
 }
